chore: refactored account migration code
